refactor(jobs): colocate per-job folders + schemaUtils
Give each scheduled job its own folder under src/jobs/, mirroring the
collection-colocation pattern from #442:

- CleanupOrphanedMedia/CleanupOrphanedMedia.ts
- SyncLectureMetadata/SyncLectureMetadata.ts

schemaUtils.ts is consumed only by CleanupOrphanedMedia, so it moves to
CleanupOrphanedMedia/schemaUtils.ts (job-internal; imported via ./schemaUtils
by the job, and directly by its test). The src/jobs/index.ts barrel still
exports only the task definitions. Purely structural; no behavior change.

Refs #444
