+++ .gitignore
@@ -22,6 +22,7 @@
 # misc
 .DS_Store
 *.pem
+*.dump
 
 # debug
 npm-debug.log*
