+++ src/jobs/SyncLectureMetadata/SyncLectureMetadata.ts
